usePomodoro 내부 코드 정리
